+++ src/packages/relay/src/__flowtests__/DeclarativeMutationConfig.js
@@ -1,14 +1,8 @@
 // @flow
 
-import { Environment, Network, RecordSource, Store } from 'relay-runtime';
+import { commitMutation, graphql, createLocalEnvironment } from '../index';
 
-import { commitMutation, graphql } from '../index';
-
-const environment = new Environment({
-  network: Network.create(() => {}),
-  store: new Store(new RecordSource()),
-});
-
+const environment = createLocalEnvironment();
 const mutation = graphql`
   mutation DeclarativeMutationConfig {
     __typename